Codegen: Parse deprecated signatures as a full FunctionSchema
Deprecated signatures are currently "parsed" manually to find the
relative order of the argument names and all other information is
inferred from the aten schema for the non-deprecated overload.
However, this leads to problems if the argument names don't match or
if there are multiple candidates that match the ATen function call.

Instead, this makes the deprecated function a full FunctionSchema and
so the entire python signature comes solely from the deprecated
schema, with the `aten:` clause only used for the dispatch lambda call.

I have confirmed locally that there is no change to
`python_torch_functionsEverything.cpp`.

ghstack-source-id: 8f4f07e8c1aa7ce527a8daec3b38ebb099c18a1c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82179